refactored the code a little to have hard-coded example data in its own module. Also changed devapi endpoints to use query parameters
