+++ client/src/services/examplecode.js
@@ -3,7 +3,7 @@ import axios from 'axios'
 const baseURL = 'http://localhost:3001'
 
 const getEx = (len) => {
-    const req = axios.get(`${baseURL}/devapi/ex/${len}`)
+    const req = axios.get(`${baseURL}/devapi/ex?length=${len}`)
     return req.then(res => res.data)
 }
