+++ server/requests/getExample.rest
@@ -1,0 +1,1 @@
+GET http://localhost:3001/devapi/ex/long
